moved handle_deprecated_data_source to data_collections
